Add cos_comm
